Lazily reset database every 5 minutes
With this change, people will be able to write to the database. The
changes will get undone every five minutes.

I debated adding pessimistic table locking to the `resets` table but
decided against it. For now, I don't think it's worth the effort, and
this solution should work well enough. 😬